# Avoid initializing properties if there is no observer. Optimize access.