--- a/tests/Avalonia.Benchmarks/Base/AvaloniaObjectInitializationBenchmark.cs
+++ b/tests/Avalonia.Benchmarks/Base/AvaloniaObjectInitializationBenchmark.cs
@@ -1,0 +1,15 @@
+﻿using Avalonia.Controls;
+using BenchmarkDotNet.Attributes;
+
+namespace Avalonia.Benchmarks.Base
+{
+    [MemoryDiagnoser]
+    public class AvaloniaObjectInitializationBenchmark
+    {
+        [Benchmark(OperationsPerInvoke = 1000)]
+        public Button InitializeButton()
+        {
+            return new Button();
+        }
+    }
+}