@MarchingCube 2019-08-30
I will run my benchmarks soon and post the results.