@ahopper 2019-08-30
This looks like a great improvement. I still wonder if AvaloniaPropertyRegistry.NotifyInitialized can be avoided completely if all it really does is set PsuedoClass defaults but that would be a far more disruptive change.