@MarchingCube 2019-08-30
Added benchmarks, will take a look at failing tests.