@MarchingCube 2019-08-30
It seems that `SelectedItems` in `TreeView` lazy functionality was broken and worked only thanks to property registry calling getter early.